+++ golden-dome/app/src/main/java/com/AliRZ/goldendome/ManualSalahClockActivity.java
@@ -1,7 +1,0 @@
-package com.AliRZ.goldendome;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class ManualSalahClockActivity extends AppCompatActivity {
-
-}