+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 07/25/2020
+- Since last update, added the manual salah clock option with all functionality such as Adhan, back buttons, location API, timezone API, Date Picker and more
+- Made a scrollable UI for the manual selection
+- Changed name of Project to Green Dome
+
 # 07/19/2020
 - Since last update, added UI Elements to the Salah Clock Activity
 - Added Functionality to most elements in the Salah Clock Activity, including Salah Timing calculated from GPS Data, Equation of Time and other equations